refactor: split workspace into standx-sdk + standx-cli crates
Extract the reusable API surface (REST client, WebSocket streams,
models, auth/signing, errors) into an independent `standx-sdk` crate
(v0.1.0) for reuse and standalone distribution. The `standx` binary
stays in `standx-cli` (v0.8.0), which re-exports the SDK surface for
backward compatibility.

- Table rendering for models is gated behind the SDK's optional
  `tabled` feature (enabled by the CLI); SDK consumers carry zero
  presentation dependencies by default
- Move the Tabled impls that lived in output.rs into the SDK (orphan
  rule) behind the same feature
- Drop unused deps: comfy-table, once_cell, config, keyring (+ the
  vestigial no-keyring feature)
- Release machinery unchanged: binary name, CI matrix, homebrew,
  install.sh all keyed on the `standx` binary

Test fixes surfaced by the split:
- Serialize env-var tests (config + credentials) with a lock; they
  raced under parallel test threads
- Wire the orphaned tests/unit/ tree into a compiled target
- test_cli_version: assert CARGO_PKG_VERSION instead of hardcoded 0.7
- order create: drop `-q` short (collides with global --quiet; clap
  debug-asserts on it), keep --qty
