+++ Cargo.toml
@@ -1,23 +1,15 @@
-[package]
-name = "standx-cli"
-version = "0.7.0"
+[workspace]
+resolver = "2"
+members = ["crates/standx-sdk", "crates/standx-cli"]
+
+[workspace.package]
 edition = "2021"
 authors = ["Kimi Claw"]
-description = "OpenClaw-first AI Agent trading toolkit - natural language to execution for any AI Agent"
 license = "MIT OR Apache-2.0"
 repository = "https://github.com/wjllance/standx-cli"
-keywords = ["openclaw", "trading", "crypto", "standx", "agent", "cli", "automation"]
-categories = ["command-line-utilities", "cryptography::cryptocurrencies", "development-tools::testing"]
 rust-version = "1.75"
 
-[[bin]]
-name = "standx"
-path = "src/main.rs"
-
-[dependencies]
-# CLI
-clap = { version = "4.5", features = ["derive", "env", "cargo"] }
-
+[workspace.dependencies]
 # Async
 tokio = { version = "1.43", features = ["full"] }
 futures = "0.3"
@@ -33,52 +25,19 @@ url = "2.5"
 # Serialization
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
-toml = "0.8"
 
 # Error handling
 anyhow = "1.0"
 thiserror = "2.0"
 
-# Configuration
-config = "0.15"
-dirs = "6.0"
-
-# Output formatting
-tabled = "0.18"
-comfy-table = "7.1"
-csv = "1.3"
-
 # Logging
 tracing = "0.1"
-tracing-subscriber = { version = "0.3", features = ["env-filter"] }
-
-# Cryptography (Ed25519)
-ed25519-dalek = { version = "2.1", features = ["rand_core"] }
-bs58 = "0.5"
-rand = "0.8"
-
-# Security
-keyring = { version = "3.0", optional = true, default-features = false }
 
 # Time
 chrono = { version = "0.4", features = ["serde"] }
 
 # Utilities
-hex = "0.4"
-base64 = "0.22"
-once_cell = "1.20"
-rpassword = "7.3"
-regex = "1.10"
-uuid = { version = "1.7", features = ["v4"] }
+dirs = "6.0"
 
-[features]
-default = []
-no-keyring = []
-
-[dev-dependencies]
-tokio-test = "0.4"
-mockito = "1.6"
-rand = "0.8"
-tempfile = "3.0"
-assert_cmd = "2.0"
-predicates = "3.0"
+# Output formatting
+tabled = "0.18"
