+++ crates/standx-cli/Cargo.toml
@@ -1,0 +1,58 @@
+[package]
+name = "standx-cli"
+version = "0.8.0"
+description = "OpenClaw-first AI Agent trading toolkit - natural language to execution for any AI Agent"
+keywords = ["openclaw", "trading", "crypto", "standx", "agent"]
+categories = ["command-line-utilities", "cryptography::cryptocurrencies"]
+edition.workspace = true
+authors.workspace = true
+license.workspace = true
+repository.workspace = true
+rust-version.workspace = true
+
+[[bin]]
+name = "standx"
+path = "src/main.rs"
+
+[dependencies]
+standx-sdk = { path = "../standx-sdk", version = "0.1.0", features = ["tabled"] }
+
+# CLI
+clap = { version = "4.5", features = ["derive", "env", "cargo"] }
+
+# Async
+tokio.workspace = true
+futures.workspace = true
+
+# Serialization
+serde.workspace = true
+serde_json.workspace = true
+toml = "0.8"
+
+# Error handling
+anyhow.workspace = true
+
+# Output formatting
+tabled.workspace = true
+csv = "1.3"
+
+# Logging
+tracing.workspace = true
+tracing-subscriber = { version = "0.3", features = ["env-filter"] }
+
+# Time
+chrono.workspace = true
+
+# Utilities
+dirs.workspace = true
+rpassword = "7.3"
+regex = "1.10"
+uuid = { version = "1.7", features = ["v4"] }
+
+[dev-dependencies]
+tokio-test = "0.4"
+mockito = "1.6"
+rand = "0.8"
+tempfile = "3.0"
+assert_cmd = "2.0"
+predicates = "3.0"
