+++ crates/standx-cli/tests/E2E_TEST_REPORT.md
