+++ crates/standx-cli/tests/TESTING_DESIGN.md
