+++ crates/standx-cli/tests/e2e/mod.rs
