+++ crates/standx-cli/tests/e2e/new_user_journey.rs
