+++ crates/standx-cli/tests/e2e/trader_workflow.rs
