+++ crates/standx-cli/tests/e2e_tests.rs
