+++ crates/standx-cli/tests/integration/cli_market_commands.rs
