+++ crates/standx-cli/tests/integration/cli_output_formats.rs
