+++ crates/standx-cli/tests/integration/mod.rs
