+++ crates/standx-cli/tests/integration_tests.rs
