+++ crates/standx-cli/tests/unit/models/market_data_test.rs
