+++ crates/standx-cli/tests/unit/models/symbol_info_test.rs
