+++ crates/standx-sdk/src/error.rs
